requirements.txt
pytest 
mysql-connector-python
